ABCN-4668: Make cache_url_opts predictable
We need to have some sort of sorting in the _cache_url_opts method, since more
than one url pattern can match. For now we trust longer patterns over shorter
patterns. If we don't sort, we trust hash order, which is random in newer
perls. We could consider adding another key in the options for weight, but lets
not complicate if we don't have to. sorting